feat: menu 생성 관련 타입 파일 이동
